requirements.txt
flask
opencv-python
numpy
ezdxf
python-occ-core
